Converting old code to function in unity

+++ Assets/Assets/Scripts/Game_Engine.cs
@@ -14,7 +14,7 @@ public class Game_Engine : MonoBehaviour
 
   private void Start()
   {
-    numberOfUnits = map.gridSizeX * map.gridSizeY / 2;
+    numberOfUnits = (int)(map.gridSizeX * map.gridSizeY / 2);
     map.numberOfUnits = numberOfUnits;
     StartGame();
   }
